+++ src/Attribute/AsTemporaryUrlServer.php
@@ -14,4 +14,4 @@ declare(strict_types=1);
 namespace Rekalogika\TemporaryUrl\Attribute;
 
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::TARGET_METHOD | \Attribute::IS_REPEATABLE)]
-class AsTemporaryUrlServer {}
+final class AsTemporaryUrlServer {}